Add ngmin as available for Clojure Ring via optimus-angular
I thought you might like to know that ngmin is integrated in an asset pipeline for Clojure. :-)

Then I noticed you had Rails listed in the readme.

I wasn't sure if you're interested in enumerating ways of using ngmin for other platforms, but if you are - here's a pull request to that effect for Clojure.